:white_check_mark: feat(users): tests added for controller/service/repo

+++ backend/src/main/java/com/collaborativeplatform/service/UserService.java
@@ -1,10 +1,12 @@
 package com.collaborativeplatform.service;
 
+import com.collaborativeplatform.exceptions.UserNotFoundException;
 import com.collaborativeplatform.model.User;
 import com.collaborativeplatform.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -19,9 +21,8 @@ public class UserService {
     return userRepository.findAll();
   }
 
-  public User getUserById(Long userId) {
-    return userRepository.findById(userId)
-      .orElseThrow(() -> new RuntimeException("User not found."));
+  public Optional<User> getUserById(Long userId) {
+    return userRepository.findById(userId);
   }
 
   public User addUser(User user) {
@@ -33,6 +34,9 @@ public class UserService {
   }
 
   public void deleteUser(Long userId) {
+    if (!userRepository.existsById(userId)) {
+      throw new UserNotFoundException("User with ID " + userId + " not found.");
+    }
     userRepository.deleteById(userId);
   }
